Create constants & move authOptions
