+++ src/app/api/admin/users/invite/route.ts
@@ -1,8 +1,8 @@
 import { getServerSession } from "next-auth/next"
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { sendInviteEmail } from "@/lib/email";
 import { prisma } from '@/lib/prisma'
 import { NextResponse } from "next/server";
+import { authOptions } from "@/constants/auth/authOptions";
 
 export async function POST(req: Request, res: Response) {
   const session = await getServerSession(authOptions)
