+++ src/app/api/auth/session/route.ts
@@ -1,7 +1,6 @@
+import { authOptions } from '@/constants/auth/authOptions'
 import { getServerSession } from 'next-auth/next'
 import { NextResponse } from 'next/server'
-import { authOptions } from '../[...nextauth]/route'
-
 
 export async function GET(request: Request) {
   const session = await getServerSession(authOptions)
