+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
 import { User } from "./user";
 import { LoginButton, LogoutButton, ThemeButton } from "@/components/buttons.components";
 import { HeroSection } from "@/components/heroSection.component";
+import { authOptions } from "@/constants/auth/authOptions";
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
